move episode list query
