--- src/common/queries/episode-list.graphql
@@ -1,18 +1,0 @@
-query EpisodeList($id: Int!, $provider: Provider!) {
-  episodes: Episodes(id: $id, provider: $provider) @client {
-    provider
-    id
-    animeId
-    title
-    duration
-    progress
-    index
-    episodeNumber
-    url
-    subtitles
-    thumbnail
-    isWatched
-  }
-}
-
-
