+++ src/graphql/documents/queries.ts
@@ -10,3 +10,22 @@ export const MEDIA_LIST_ENTRY_FROM_MEDIA_ID = gql`
 
   ${ANILIST_LIST_ENTRY_FRAGMENT}
 `
+
+export const EPISODE_LIST = gql`
+  query EpisodeList($id: Int!, $provider: Provider!) {
+    episodes: Episodes(id: $id, provider: $provider) @client {
+      provider
+      id
+      animeId
+      title
+      duration
+      progress
+      index
+      episodeNumber
+      url
+      subtitles
+      thumbnail
+      isWatched
+    }
+  }
+`
