+++ src/views/queue/components/queue-item.vue
@@ -123,7 +123,7 @@ import { Component, Vue } from 'vue-property-decorator'
 import { oc } from 'ts-optchain'
 import { mdiChevronDown, mdiMenu } from '@mdi/js'
 
-import EPISODE_LIST from '@/common/queries/episode-list.graphql'
+import { EPISODE_LIST } from "@/graphql/documents/queries"
 import { setProgress } from '@/common/mutations/episodes'
 import { startRewatching, updateStatus } from '@/common/mutations/list-entry'
 import {
